[Compression] Rename package (compress -> compression) to be consistent with fs2/zio
