+++ monix-reactive/jvm/src/test/scala/monix/reactive/compression/InflateTest.scala
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package monix.reactive.compress
+package monix.reactive.compression
 
 import monix.reactive.Observable
